+++ StreamVideoTests/Mock/MockCall.swift
@@ -8,10 +8,12 @@ import Foundation
 final class MockCall: Call, Mockable, @unchecked Sendable {
 
     typealias FunctionKey = MockCallFunctionKey
+    typealias FunctionInputKey = MockCallFunctionInputKey
 
     enum MockCallFunctionKey: Hashable, CaseIterable {
         case get
         case accept
+        case reject
         case join
         case updateTrackSize
         case callKitActivated
@@ -30,6 +32,8 @@ final class MockCall: Call, Mockable, @unchecked Sendable {
 
         case callKitActivated(audioSession: AVAudioSessionProtocol)
 
+        case reject(reason: String?)
+
         var payload: Any {
             switch self {
             case let .join(create, options, ring, notify, callSettings):
@@ -40,15 +44,17 @@ final class MockCall: Call, Mockable, @unchecked Sendable {
 
             case let .callKitActivated(audioSession):
                 return audioSession
+
+            case let .reject(reason):
+                return reason ?? ""
             }
         }
     }
 
     var stubbedProperty: [String: Any]
     var stubbedFunction: [FunctionKey: Any] = [:]
-    @Atomic var stubbedFunctionInput: [FunctionKey: [MockCallFunctionInputKey]] = MockCallFunctionKey
-        .allCases
-        .reduce(into: [FunctionKey: [MockCallFunctionInputKey]]()) { $0[$1] = [] }
+    @Atomic var stubbedFunctionInput: [FunctionKey: [FunctionInputKey]] = FunctionKey.allCases
+        .reduce(into: [FunctionKey: [FunctionInputKey]]()) { $0[$1] = [] }
 
     override var state: CallState {
         get { self[dynamicMember: \.state] }
@@ -89,6 +95,11 @@ final class MockCall: Call, Mockable, @unchecked Sendable {
 
     override func accept() async throws -> AcceptCallResponse {
         stubbedFunction[.accept] as! AcceptCallResponse
+    }
+
+    override func reject(reason: String? = nil) async throws -> RejectCallResponse {
+        stubbedFunctionInput[.reject]?.append(.reject(reason: reason))
+        return stubbedFunction[.reject] as! RejectCallResponse
     }
 
     override func join(