+++ StreamVideoTests/WebRTC/v2/WebRTCStateAdapter_Tests.swift
@@ -64,6 +64,16 @@ final class WebRTCStateAdapter_Tests: XCTestCase, @unchecked Sendable {
         await assertEqualAsync(await subject.sessionID, expected)
     }
 
+    // MARK: - setIsTracingEnabled
+
+    func test_setIsTracingEnabled_shouldUpdateIsTracingEnabled() async throws {
+        let expected = true
+
+        await subject.set(isTracingEnabled: expected)
+
+        await assertEqualAsync(await subject.isTracingEnabled, expected)
+    }
+
     // MARK: - setCallSettings
 
     func test_setCallSettings_shouldUpdateCallSettings() async throws {
@@ -164,27 +174,38 @@ final class WebRTCStateAdapter_Tests: XCTestCase, @unchecked Sendable {
         await assertEqualAsync(await subject.ownCapabilities, expected)
     }
 
-    // MARK: - setStatsReporter
+    // MARK: - setStatsAdapter
 
-    func test_setStatsReporter_shouldUpdateStatsReporter() async throws {
-        let expected = WebRTCStatsReporter(sessionID: .unique)
+    func test_setStatsReporter_shouldUpdateStatsAdapter() async throws {
+        let expected = WebRTCStatsAdapter(
+            sessionID: .unique,
+            unifiedSessionID: .unique,
+            isTracingEnabled: true,
+            trackStorage: await subject.trackStorage
+        )
 
-        await subject.set(statsReporter: expected)
+        await subject.set(statsAdapter: expected)
 
-        await assertTrueAsync(await subject.statsReporter === expected)
+        await assertTrueAsync(await subject.statsAdapter === expected)
+        await assertTrueAsync(await subject.audioSession === expected.audioSession)
     }
 
     // MARK: - setSFUAdapter
 
     func test_setSFUAdapter_shouldUpdateSFUAdapterAndStatsReporter() async throws {
-        let statsReporter = WebRTCStatsReporter(sessionID: .unique)
-        await subject.set(statsReporter: statsReporter)
+        let statsAdapter = WebRTCStatsAdapter(
+            sessionID: .unique,
+            unifiedSessionID: .unique,
+            isTracingEnabled: true,
+            trackStorage: await subject.trackStorage
+        )
+        await subject.set(statsAdapter: statsAdapter)
         let mockSFUStack = MockSFUStack()
 
         await subject.set(sfuAdapter: mockSFUStack.adapter)
 
         await assertTrueAsync(await subject.sfuAdapter === mockSFUStack.adapter)
-        XCTAssertTrue(statsReporter.sfuAdapter === mockSFUStack.adapter)
+        XCTAssertTrue(statsAdapter.sfuAdapter === mockSFUStack.adapter)
     }
 
     // MARK: - setParticipantsCount
@@ -429,7 +450,7 @@ final class WebRTCStateAdapter_Tests: XCTestCase, @unchecked Sendable {
         await fulfillment { await self.subject.publisher == nil }
         await assertNilAsync(await subject.publisher)
         await assertNilAsync(await subject.subscriber)
-        await assertNilAsync(await subject.statsReporter)
+        await assertNilAsync(await subject.statsAdapter)
         await assertNilAsync(await subject.sfuAdapter)
         await assertEqualAsync(await subject.token, "")
         await assertEqualAsync(await subject.sessionID, "")
@@ -505,7 +526,7 @@ final class WebRTCStateAdapter_Tests: XCTestCase, @unchecked Sendable {
         XCTAssertEqual(sfuStack.webSocket.timesCalled(.disconnectAsync), 0)
         await assertNilAsync(await subject.publisher)
         await assertNilAsync(await subject.subscriber)
-        await assertNilAsync(await subject.statsReporter)
+        await assertNilAsync(await subject.statsAdapter)
         await assertNilAsync(await subject.sfuAdapter)
         await assertEqualAsync(await subject.token, "")
         await assertEqualAsync(await subject.sessionID, sessionId)
@@ -534,7 +555,6 @@ final class WebRTCStateAdapter_Tests: XCTestCase, @unchecked Sendable {
         )
         await subject.set(callSettings: .init(cameraPosition: .back))
 
-        let sessionId = await subject.sessionID
         await subject.cleanUpForReconnection()
 
         await assertEqualAsync(await subject.callSettings.cameraPosition, .back)
@@ -889,7 +909,14 @@ final class WebRTCStateAdapter_Tests: XCTestCase, @unchecked Sendable {
         await subject.set(participantPins: participantPins)
         await subject.enqueue { _ in participants }
         try await subject.configurePeerConnections()
-        await subject.set(statsReporter: WebRTCStatsReporter(sessionID: .unique))
+
+        let statsAdapter = WebRTCStatsAdapter(
+            sessionID: .unique,
+            unifiedSessionID: .unique,
+            isTracingEnabled: true,
+            trackStorage: await subject.trackStorage
+        )
+        await subject.set(statsAdapter: statsAdapter)
     }
 }